requirements.txt
kafka
kafka-python
kafka-python-ng
psycopg2
python-dotenv
pandas
pyarrow
bson
minio
PyYAML
s3fs==0.4.2
delta-spark==3.2.1
google-cloud-bigquery
deltalake
pyspark==3.5.3